inter-model communication pipeline/pyproject.toml
[tool.poetry]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = ">=2.5"
pandas = "^2.2"
python-dotenv = "^1.0.1"
diffusers = "^0.32.0"
transformers = "^4.48"
safetensors = "^0.4.5"
tqdm = "^4.66.6"
huggingface-hub = {extras = ["cli"], version = "^0.26.2"}
aiohttp = "^3.10.10"
sentence-transformers = "^3.4.1"
argparse = "^1.4.0"
scipy = "^1.9"
accelerate = "^1"
pyyaml = "^6.0.2"
pydantic = "^2.9.2"
protobuf = "^5.28.3"
sentencepiece = "^0.2.0"
matplotlib = "^3.7"
adjusttext = "^1.3.0"
krippendorff = "0.6.0"
gdown = "^5.2.0"
statsmodels = "^0.14.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5"
gdown = "^5.2.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
